requirements.txt
pandas
ccxt
mysql-connector-python
psycopg2-binary
SQLAlchemy
ta
bybit
python-dotenv
pymysql
pybit
pycryptodome
pandas_ta
joblib
tensorflow-cpu
numpy
DateTime
yfinance
requests
scikit-learn
pytrends
psycopg2
plotly
nbformat